requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
praw==6.3.1
prawcore==1.0.1
pycrypto==2.6.1
pyTelegramBotAPI==3.6.6
requests==2.22.0
six==1.12.0
update-checker==0.16
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.6
